created an hash map

+++ 0x1A-hash_tables/6-hash_table_delete.c
@@ -5,7 +5,7 @@
  */
 void hash_table_delete(hash_table_t *ht)
 {
-	unsigned int i = 0;
+	unsigned long int i = 0;
 	hash_node_t *node = NULL;
 	hash_node_t *tmp = NULL;
